pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13"
tzdata = "^2025.2"
psycopg = {extras = ["binary", "pool"], version = "^3.2.9"}
gunicorn = "23.0.0"
pillow = "11.3.0"
django = "5.2.5"
libsass = "0.23.0"
django-sass-processor = "1.4.2"
django-compressor = "4.5.1"
djangorestframework = "3.16.1"
django-filter = "25.1"
django-phonenumber-field = { extras = ["phonenumberslite"], version = "8.1.0" }
python-decouple = "3.8"
django-ckeditor-5 = "0.2.18"

[tool.poetry.group.dev.dependencies]
djlint = "^1.36.4"

[tool.djlint]
format_css = true
ignore = "H006,H016,H021,H023,H030,H031,T002"
blank_line_before_tag = "if,block,for"
blank_line_after_tag = "load,endif,endblock,endfor"

[tool.djlint.css]
indent_size = 2

[tool.ruff]
line-length = 101